refactor(ds): a portalled listbox popup anchors one way, not two
Select and the combo box about to join it open the same popup over
different triggers, and every rule in here is about the popup rather
than about what opened it.
